P2022-1557 Working common validation
